+++ nautilus_f12/config.py
@@ -1,0 +1,24 @@
+"""
+Nautilus F12 Configuration & Constants
+"""
+
+DEFAULT_TERMINAL_HEIGHT = 240
+DEFAULT_FONT = "Monospace 10"
+DEFAULT_TOGGLE_KEY = "F12"
+WORD_CHAR_EXCEPTIONS = "-#%&+,./:=?@_~"
+
+EXPAND_VIEW_WIDGET_NAMES = [
+    "GtkOverlay",
+    "NautilusCanvasView",
+    "NautilusViewIconController",
+    "NautilusListView",
+    "NautilusFilesView",
+    "GtkScrolledWindow",
+    "AdwToastOverlay",
+]
+
+SLOT_CONTAINER_NAMES = [
+    "NautilusWindowSlot",
+    "NautilusWindowSlotView",
+    "GtkBox",
+]