+++ nautilus_f12/input/__init__.py
@@ -1,0 +1,3 @@
+from .key_controller import WindowKeyController
+
+__all__ = ["WindowKeyController"]